% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{bs_restore_tabs}
\alias{bs_restore_tabs}
\title{Restore Tabs}
\usage{
bs_restore_tabs(config, input, output, session)
}
\arguments{
\item{config}{A config object.}

\item{input, output, session}{Shiny parameters.}
}
\description{
Restore tabs from a config object.
}
