% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{init_conf}
\alias{init_conf}
\title{Store loaded config}
\usage{
init_conf(conf)
}
\arguments{
\item{conf}{The config object.}
}
\description{
Store the initially loaded config in the storage environment.
}
\keyword{internal}
