% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{json}
\alias{json}
\alias{save_json}
\alias{get_json}
\title{Save the config to a JSON file}
\usage{
save_json(env, input, output, session)

get_json()
}
\arguments{
\item{env}{The config environment.}

\item{input}{The input object.}

\item{output}{The output object.}

\item{session}{The session object.}
}
\description{
Save the config to a JSON file.
}
