% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{parse_blockr}
\alias{parse_blockr}
\title{Parse serialised worksplace}
\usage{
parse_blockr(conf, ...)
}
\arguments{
\item{conf}{The config object.}

\item{...}{Ignored.}
}
\description{
Parse serialised worksplace and store in environment.
}
