% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{save}
\alias{save}
\alias{with_block_app}
\alias{block_app}
\title{blockApp}
\usage{
with_block_app(app, get_config = get_json, restore_tabs = bs_restore_tabs)

block_app(
  ui,
  server,
  get_config = get_json,
  restore_tabs = bs_restore_tabs,
  ...
)
}
\arguments{
\item{app}{App object.}

\item{get_config}{Function that returns the configuration.}

\item{restore_tabs}{Function that restores the tabs.}

\item{ui}{App UI function.}

\item{server}{App server function.}

\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
An application that uses blocks and will restore state.
}
