% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{set_masonry}
\alias{set_masonry}
\title{Set masonry config}
\usage{
set_masonry(tabid, conf)
}
\arguments{
\item{tabid}{ID of tab.}

\item{conf}{Masonry config.}
}
\description{
Set masonry config for a tab in the storage environment.
}
