% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{set_tab}
\alias{set_tab}
\title{Set tab data}
\usage{
set_tab(label, ..., content = list(), id = label, custom = list())
}
\arguments{
\item{label}{The tab's label.}

\item{...}{The tab's stacks.}

\item{content}{The tab's content, other than stacks.}

\item{id}{ID of tab.}

\item{custom}{Custom tab data to use with a custom restore function.}
}
\description{
Set a tab's data in the storage environment.
}
