% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{save}
\alias{save}
\alias{with_blockr_app}
\alias{blockr_app}
\title{blockApp}
\usage{
with_blockr_app(
  app,
  save_config = save_json,
  get_config = get_json,
  restore_tabs = bs_restore_tabs,
  custom = NULL
)

blockr_app(
  ui,
  server,
  save_config = save_json,
  get_config = get_json,
  restore_tabs = bs_restore_tabs,
  custom = NULL,
  ...
)
}
\arguments{
\item{app}{App object.}

\item{save_config}{Function that saves the configuration.}

\item{get_config}{Function that returns the configuration.}

\item{restore_tabs}{Function that restores the tabs.}

\item{custom}{Cumstom function to run after restoring tabs.
Must accept \code{config}, \code{input}, \code{output}, and \code{session} arguments.}

\item{ui}{App UI function.}

\item{server}{App server function.}

\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
An application that uses blocks and will restore state.
}
\keyword{internal}
